+++ src/main/java/com/project/catxi/common/auth/controller/OAuthController.java
@@ -1,16 +1,13 @@
 package com.project.catxi.common.auth.controller;
 
 import com.project.catxi.common.api.ApiResponse;
-import com.project.catxi.common.auth.kakao.BlacklistDTO;
 import com.project.catxi.common.auth.kakao.KakaoDTO;
 import com.project.catxi.common.auth.kakao.TokenDTO;
-import com.project.catxi.common.auth.service.BlacklistService;
 import com.project.catxi.common.auth.service.OAuthLoginService;
 import com.project.catxi.common.auth.service.TokenService;
 import com.project.catxi.common.domain.MemberStatus;
 import com.project.catxi.member.domain.Member;
 import com.project.catxi.member.dto.CustomUserDetails;
-import com.project.catxi.member.repository.MemberRepository;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -33,8 +30,6 @@ public class OAuthController {
 
   private final OAuthLoginService oAuthLoginService;
   private final TokenService tokenService;
-  private final BlacklistService blacklistService;
-  private final MemberRepository memberRepository;
 
   @Operation(summary = "인가코드 로그인", description = "카카오에서 인가코드를 받아와 JWT를 발급해줍니다.")
   @GetMapping("/kakao/callback")
@@ -91,41 +86,6 @@ public class OAuthController {
 
     tokenService.logout(request, refreshToken, response);
     return ApiResponse.success("로그아웃 완료");
-  }
-
-  //블랙리스트 등록
-  @Operation(summary = "사용자 ID로 블랙리스트 등록", description = "사용자를 블랙리스트에 등록합니다. durationDays가 null이면 영구 등록됩니다.")
-  @PostMapping("/uID/{userId}")
-  public ApiResponse<?> addUserToBlacklistById(@PathVariable Long userId) {
-
-    blacklistService.addUserToBlacklistPermanent(userId);
-    return ApiResponse.success("사용자가 블랙리스트에 등록되었습니다.");
-  }
-
-  //블랙리스트 해제
-  @Operation(summary = "사용자 ID로 블랙리스트 해제", description = "사용자를 블랙리스트에서 해제합니다.")
-  @DeleteMapping("/uID/{userId}")
-  public ApiResponse<?> removeUserFromBlacklistById(@PathVariable Long userId) {
-
-    blacklistService.removeUserFromBlacklist(userId);
-    return ApiResponse.success("사용자가 블랙리스트에서 해제되었습니다.");
-  }
-
-  //블랙리스트 여부 조회
-  @Operation(summary = "사용자 ID로 블랙리스트 상태 확인", description = "사용자가 블랙리스트에 등록되어 있는지 확인합니다.")
-  @GetMapping("/blacklist/uID/{userId}")
-  public ApiResponse<BlacklistDTO.checkBlacklist> checkBlacklistStatusById(@PathVariable Long userId) {
-
-    boolean isBL = blacklistService.isUserBlacklisted(userId);
-
-    Member member = memberRepository.findById(userId)
-        .orElse(null);
-    String email = member != null ? member.getEmail() : "Unknown";
-
-    BlacklistDTO.checkBlacklist response
-        = new BlacklistDTO.checkBlacklist(isBL, userId, email);
-
-    return ApiResponse.success(response);
   }
 
   //TODO: 회원 탈퇴
